+++ es7/promise.js
@@ -1,0 +1,3 @@
+require('../modules/es6.promise');
+require('../modules/es7.promise.finally');
+module.exports = require('../modules/_core').Promise;