+++ library/es7/promise.js
@@ -1,6 +1,3 @@
-require('../modules/es6.object.to-string');
-require('../modules/es6.string.iterator');
-require('../modules/web.dom.iterable');
 require('../modules/es6.promise');
 require('../modules/es7.promise.finally');
 module.exports = require('../modules/_core').Promise;