+++ library/fn/promise/finally.js
@@ -1,0 +1,2 @@
+require('../../modules/es7.promise.finally');
+module.exports = require('../../modules/_core').Promise.finally;